+++ src/lightning/fabric/plugins/environments/slurm.py
@@ -141,6 +141,22 @@ class SLURMEnvironment(ClusterEnvironment):
     def node_rank(self) -> int:
         return int(os.environ["SLURM_NODEID"])
 
+    def validate_settings(self, num_devices: int, num_nodes: int) -> None:
+        if _is_slurm_interactive_mode():
+            return
+        ntasks_per_node = os.environ.get("SLURM_NTASKS_PER_NODE")
+        if ntasks_per_node is not None and int(ntasks_per_node) != num_devices:
+            raise ValueError(
+                f"You set `devices={num_devices}` in Lightning, but the number of tasks per node configured in SLURM"
+                f" `--ntasks-per-node={ntasks_per_node}` does not match. HINT: Set `devices={ntasks_per_node}`."
+            )
+        nnodes = os.environ.get("SLURM_NNODES")
+        if nnodes is not None and int(nnodes) != num_nodes:
+            raise ValueError(
+                f"You set `num_nodes={num_nodes}` in Lightning, but the number of nodes configured in SLURM"
+                f" `--nodes={nnodes}` does not match. HINT: Set `num_nodes={nnodes}`."
+            )
+
     @staticmethod
     def resolve_root_node_address(nodes: str) -> str:
         """The node selection format in SLURM supports several formats.
@@ -182,8 +198,7 @@ class SLURMEnvironment(ClusterEnvironment):
         """Checks for conflicting or incorrectly set variables set through `srun` and raises a useful error message.
 
         Right now, we only check for the most common user errors. See
-        `the srun docs <https://slurm.schedmd.com/srun.html>`_
-        for a complete list of supported srun variables.
+        `the srun docs <https://slurm.schedmd.com/srun.html>`_ for a complete list of supported srun variables.
 
         """
         ntasks = int(os.environ.get("SLURM_NTASKS", "1"))